removed wrong error case

+++ binaire_macho/diff_nm_linux.sh
@@ -3,9 +3,14 @@ rm errftnm errnm
 for file in $@ $1
 do
 	echo "diff on $file"
-	echo $file >> errftnm
-	cat $file.out > resnm 2>> errnm
-	../nm/ft_nm $file > resftnm 2>> errftnm
-	diff resnm resftnm
+	if [ -s "$file.err" ]
+	then
+		echo "err by nm"
+	else
+		echo $file >> errftnm
+		cat $file.out > resnm 2>> errnm
+		../nm/ft_nm $file > resftnm 2>> errftnm
+		diff resnm resftnm
+	fi
 done
 rm resnm resftnm
